refactor(cli): group helper image commands into image sub-package
