+++ cmd/internal/helper/image/get_image.go
@@ -1,4 +1,4 @@
-package helper
+package image
 
 import (
 	"context"
